Add functionality to calculator
To make the calculator starts working like a real one, there are
four global variables, two for the numbers that will be used to make
an operation, another one for the operator that was pressed and a
boolean that tells if the user has pressed a new number or not.

Also there are changes to the functions that make the operations. Now
they return the value of that operation, which will be used to display
on the screen and store it if the user wants to make another operation.

The clear button now deletes the global variables values and displays a
0 (the default value on the screen) to avoid some bugs in case the user
presses another unexpected button.

The digit buttons now first verify if the user has pressed a digit
button before or not, in case it's the first digit button pressed, it
clears the screen and changes the value of the numberPressed variable to
indicate that the user is pressing a new number.

activateOperatorButtons() function is where it's the logic of the
operations. First, it declares a variable to store the result of an
operation, then adds functionality to the buttons with the
eventListener.

In the function that executes when a user presses an operator button,
verifies if the user has captured a new number, if it is, then it gets
the dispĺay's value, then verifies with the global variable if there is
a number stored, in case it's not, it stores as the first number and the
operator pressed. When there is a number stored, it stores the second
number, clear the screen, do the operation and update the screen with
that result, store that result as the first number if the user wants
to make another operation, delete the second number and finally indicate
that the user has not captured another number.