Add functionality to equals <button>
First, the float numbers are now rounded to 2 decimals.

The equals <button> now has a functionality similar to the operators
with the eventListener inside the new activateEqualsButton() function.

First, checks if the user captured a number, if it is then check if
there are two numbers and one operator pressed, if it is then makes
the operation, stores it in a variable, store it as the first number
in case the user makes another operation, displays the result on the
display, deletes the operator to update it and finally marks with
numberPressed that the user can press another number to do a new
operation.