Fix bug about consecutive operations
When the user makes an operation and then pressed the equals button,
it didn't delete the second number and the operator pressed and that
causes to make the wrong operations.

Now the calculator can do consecutive actions if the user press the
equals button and wants to make another operation(s).